Added Dumper.quote()
Return a representation of the object in a format useful for embedding
in SQL. For most objects this is only the representation in quotes and
escaped, for some it's a different value (numbers, bools, NULL).

None can be quoted but not dumped, so it can be passed to an
sql.Literal() without the need of a special case.