Add AsyncRowFactory class
The extra class allow clients to define their RowFactory just taking a
Cursor or an AsyncCursor, which is easier if the program handles only
one type of connection (sync or async).

Using a Generic, the server-side cursors are now subclassed from  the
respective client-side cursor (sync and async), which allows to drop a
bit of implementation duplication.