Replace git_row_types() with get_dumper_types/get_loader_types()
dump_sequence() can make use of the dumpers pre-set by get_dumper_types()
and is now used in composite binary dump and in binary COPY FROM.

The interface should be better iterated because for the latter use cases
the extra info (oids, formats) are just a waste of resources.

Only the Python implementation has been changed so far, the C
implementation will be changed down the line.