Plug high-level result retrieval into pipeline mode
We add a _pipeline_queue attribute on connection object that is used to
keep track of queries sent in a pipeline. Thus the queue is filled every
time we call send() generator. If sent query comes from a
cursor.execute() call, we append respective cursor to the queue so as to
later bind results from that query to the cursor. Otherwise, e.g. in
case of a BEGIN command or a PQpipelineSync() call, we just append a
None value to the queue since we do not need to attach results to a
cursor upon retrieval.

Upon results fetch (cursor.fetch*() method call), we call the new
_fetch_pipeline() method which delegate to BaseCursor's
_fetch_pipeline_gen() generator; the later will repeatedly fetch results
from the pipeline until the cursor at stake is found in the queue. For
each cursor found in the queue while processing, results are attached
(and the cursor reset).

In Cursor._raise_from_results() we now handle the PIPELINE_ABORTED
result status that is converted into an OperationalError.

At pipeline exit (in the pipeline() context manager), we process
pending results, possibly raising a ProgrammingError if any of those
concern a cursor that the user should have fetched results from; this is
implemented as BaseConnection._end_pipeline_gen().