Make writable pipeline_status a property of PGconn
This property replaces enter_pipeline_mode() and exit_pipeline_mode()
and is now used in Connection.pipeline() to enter/exit the pipeline
mode. The 'status' method of Pipeline is now a property.

Accordingly, the _pipeline_mode attribute on Connection is dropped and
pgconn.pipeline_status is used instead. In order to keep compatibility
with libpq older than 14.0, we always return False from
pgconn.pipeline_status in that case.

Note that we have to allow both bool and PipelineStatus value for the
pipeline_status property in order to please mypy, though we'd only ever
return a PipelineStatus value.