Move pipeline results processing method on BaseConnection
We move _fetch_pipeline_gen() on BaseConnection and use it upon exit of
the pipeline in place of slightly different _end_pipeline_gen(). From
now on, the logic for results processing in the middle of the pipeline
or at exit is similar. The only difference is that we still only fetch
results as needed upon cursor.fetch*(), meaning we stop once the cursor
at stake is found in the queue.