Wait for prior results in execute() in pipeline mode
In pipeline mode, we now have cursor.execute() to possibly block, using
added '_queued' Event attribute, waiting on results from a possible prior
execute() to be fetched. The event is clear()ed at the end of execute()
and then set() upon entering a fetch*() call, through _fetch_pipeline(),
when results are set.

In pipeline tests, in order to avoid the new locks, we now avoid
consecutive execute() on a single cursor by using distinct cursors. To
demonstrate blocking mechanism, test_pipeline_execute_wait() is added
with threading or asyncio.

The '# type: ignore[misc]' in AsyncCursor.execute() is needed because of
'self' being declared as AnyCursor.