Possibly send a Ready.RW value to generators waiting on RW
In some generators, we might be interested in receiving both read-ready
and write-ready event at the same time.

The previous implementation of wait_epoll() and wait_async() did not
allow that: wait_epoll() would return Ready.R, even if the event was
'EPOLLIN | EPOLLOUT', and wait_async() would just send the first ready
value (because of the previous Event-based strategy).

On the other hande, wait_selector() already worked fine, by accident
since it did not return a true Ready value, but an integer; this is
fixed now.

In tests, this is demonstrated by using a connection in pipeline mode,
with a "large" query and a little sleep.