Make test_range.py mypy-clean
This consists in adding the type variable to Range type when it makes
sense or to let mypy ignore the 'var-annotated' error for ranges
declared as empty, and where we do not use the underlying value.

In test_in(), we add an extra check to verify that mypy warns about
types incompatibility.