Schedule a daily test run and a weekly package run
This should help detecting breakages due to new depending packages. In
the future we may extend the tests to run on Python/Postgres master
branch too.

The intention is to remove the top bound from dependency packages,
because the practice is widely documented to be problematic. Thankfully
we have only dev/test dependencies, but I'd rather get the habit for
good.

This commit will be merged on master as it is because otherwise Github
won't pick the change.