feat: wrap transaction in pipelines if the connection has one
When the connection is in pipeline mode, we wrap each transaction block
with two pipelines: the inner one ensures that we'd get a Sync at the
end of transaction *before* the commit or rollback statement, thus
ensuring that the latter gets run even if the pipeline got in aborted
state during the transaction; the outer one ensures (also emitting a
sync) that the connection is restored into a usable state even when the
transaction rolled back.

Introduce nullcontext and Abstract(Async)ContextManager compat
definitions.